requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
bokeh==3.7.0
    # via -r requirements.in
build==1.2.2.post1
    # via
    #   -r requirements.in
    #   pip-tools
click==8.1.8
    # via
    #   -r requirements.in
    #   pip-tools
contourpy==1.3.1
    # via
    #   -r requirements.in
    #   bokeh
    #   matplotlib
cycler==0.12.1
    # via
    #   -r requirements.in
    #   matplotlib
filelock==3.18.0
    # via
    #   -r requirements.in
    #   torch
fonttools==4.56.0
    # via
    #   -r requirements.in
    #   matplotlib
fsspec==2025.3.0
    # via
    #   -r requirements.in
    #   torch
jinja2==3.1.6
    # via
    #   -r requirements.in
    #   bokeh
    #   torch
joblib==1.4.2
    # via
    #   -r requirements.in
    #   scikit-learn
kiwisolver==1.4.8
    # via
    #   -r requirements.in
    #   matplotlib
markupsafe==3.0.2
    # via
    #   -r requirements.in
    #   jinja2
matplotlib==3.10.1
    # via
    #   -r requirements.in
    #   seaborn
mpmath==1.3.0
    # via
    #   -r requirements.in
    #   sympy
narwhals==1.31.0
    # via
    #   -r requirements.in
    #   bokeh
networkx==3.4.2
    # via
    #   -r requirements.in
    #   torch
numpy==2.2.4
    # via
    #   -r requirements.in
    #   bokeh
    #   contourpy
    #   matplotlib
    #   pandas
    #   scikit-learn
    #   scipy
    #   seaborn
    #   torchvision
nvidia-cublas-cu12==12.4.5.8
    # via
    #   -r requirements.in
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.4.127
    # via
    #   -r requirements.in
    #   torch
nvidia-cuda-nvrtc-cu12==12.4.127
    # via
    #   -r requirements.in
    #   torch
nvidia-cuda-runtime-cu12==12.4.127
    # via
    #   -r requirements.in
    #   torch
nvidia-cudnn-cu12==9.1.0.70
    # via
    #   -r requirements.in
    #   torch
nvidia-cufft-cu12==11.2.1.3
    # via
    #   -r requirements.in
    #   torch
nvidia-curand-cu12==10.3.5.147
    # via
    #   -r requirements.in
    #   torch
nvidia-cusolver-cu12==11.6.1.9
    # via
    #   -r requirements.in
    #   torch
nvidia-cusparse-cu12==12.3.1.170
    # via
    #   -r requirements.in
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.6.2
    # via
    #   -r requirements.in
    #   torch
nvidia-nccl-cu12==2.21.5
    # via
    #   -r requirements.in
    #   torch
nvidia-nvjitlink-cu12==12.4.127
    # via
    #   -r requirements.in
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.4.127
    # via
    #   -r requirements.in
    #   torch
packaging==24.2
    # via
    #   -r requirements.in
    #   bokeh
    #   build
    #   matplotlib
pandas==2.2.3
    # via
    #   -r requirements.in
    #   bokeh
    #   seaborn
pillow==11.1.0
    # via
    #   -r requirements.in
    #   bokeh
    #   matplotlib
    #   torchvision
pip-tools==7.4.1
    # via -r requirements.in
plotly==5.13.0
    # via -r requirements.in
pyparsing==3.2.1
    # via
    #   -r requirements.in
    #   matplotlib
pyproject-hooks==1.2.0
    # via
    #   -r requirements.in
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via
    #   -r requirements.in
    #   matplotlib
    #   pandas
pytz==2025.1
    # via
    #   -r requirements.in
    #   pandas
pyyaml==6.0.2
    # via
    #   -r requirements.in
    #   bokeh
scikit-learn==1.6.1
    # via -r requirements.in
scipy==1.15.2
    # via
    #   -r requirements.in
    #   scikit-learn
seaborn==0.13.2
    # via -r requirements.in
six==1.17.0
    # via
    #   -r requirements.in
    #   python-dateutil
sympy==1.13.1
    # via
    #   -r requirements.in
    #   torch
tenacity==9.0.0
    # via plotly
threadpoolctl==3.6.0
    # via
    #   -r requirements.in
    #   scikit-learn
tomli==2.2.1
    # via
    #   -r requirements.in
    #   build
    #   pip-tools
torch==2.6.0
    # via
    #   -r requirements.in
    #   torchvision
torchvision==0.21.0
    # via -r requirements.in
tornado==6.4.2
    # via
    #   -r requirements.in
    #   bokeh
triton==3.2.0
    # via
    #   -r requirements.in
    #   torch
typing-extensions==4.12.2
    # via
    #   -r requirements.in
    #   torch
tzdata==2025.1
    # via
    #   -r requirements.in
    #   pandas
wheel==0.45.1
    # via pip-tools
xyzservices==2025.1.0
    # via
    #   -r requirements.in
    #   bokeh

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
